+++ api-server/src/test/kotlin/com/unicorn/api/controller/greeting/GreetingsGetTest.kt
@@ -1,6 +1,5 @@
-package com.unicorn.api.controller
+package com.unicorn.api.controller.greeting
 
-import com.unicorn.api.query_service.GreetingDto
 import org.junit.jupiter.api.Test
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc
@@ -11,38 +10,30 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.content
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers.status
 import org.springframework.transaction.annotation.Transactional
-import java.util.*
 
 @ActiveProfiles("test")
 @SpringBootTest
 @AutoConfigureMockMvc
 @Transactional
-class GreetingGetTest {
+class GreetingsGetTest {
+
     @Autowired
     private lateinit var mockMvc: MockMvc
 
     @Test
     fun `should return 200 when greeting is called`() {
-        val greeting = GreetingDto(
-            id = UUID.fromString("f47ac10b-58cc-4372-a567-0e02b2c3d479"),
-            message = "Hello, World!"
-        )
-        val result = mockMvc.perform(MockMvcRequestBuilders.get("/greetings/${greeting.id}"))
-
+        val result = mockMvc.perform(MockMvcRequestBuilders.get("/greetings"))
 
         result.andExpect(status().isOk)
         result.andExpect(content().json("""
             {
-                "id": "${greeting.id}",
-                "message": "${greeting.message}"
+                "greetings": [
+                    {
+                        "id": "f47ac10b-58cc-4372-a567-0e02b2c3d479",
+                        "message": "Hello, World!"
+                    }
+                ]
             }
         """.trimIndent(), true))
-    }
-
-    @Test
-    fun `should return 404 when greeting is not found`() {
-        val result = mockMvc.perform(MockMvcRequestBuilders.get("/greetings/${UUID.randomUUID()}"))
-
-        result.andExpect(status().isNotFound)
     }
 }
